Lot of bug fixed, maybe ready for final release -_-

+++ whynotimagecarousel/src/main/java/org/imaginativeworld/whynotimagecarousel/Extensions.kt
@@ -15,8 +15,8 @@ import androidx.recyclerview.widget.SnapHelper
  * @param context Context to get resources and device specific display metrics
  * @return A float value to represent dp equivalent to px value
  */
-fun Float.toDp(context: Context): Float {
-    return this / (context.resources.displayMetrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)
+fun Int.toDp(context: Context): Int {
+    return (this / (context.resources.displayMetrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)).toInt()
 }
 
 /**
